Create experiments.tsv
This is draft 1 of the wgEncode to ENCSR mapping.  There is a notes column with three categories of mappings.  

NOT IN MDB: These are experiments that may or may not have an ENCSR number made for it where no data showed up in the UCSC mdb.

ONE TO MANY:  These are experiments (wgEncode) that have broken into many ENCSR's.  They will have their own file-by-file mapping in this repository.

analysis dataset:  This should not really affect things, but is nice to know.  These are datasets where there was no wetlab experiment associated, but rather lots of analysis files.  They are given an ENCSR of type project rather than type experiment.

There are MANY-TO-ONE situations in this mapping, but I did not note them as a problem.  Maybe in version 2 I will note those.